requirements.txt
requests==2.31.0
python-dotenv==1.0.0
sentry-sdk==2.27.0
pyairtable==1.4.0
Pillow==10.0.0
aiohttp==3.8.5
aiofiles==24.1.0
openai==1.65.5
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-asyncio==0.21.1
httpx==0.24.1
beautifulsoup4==4.12.2
tenacity==8.2.3
pydantic==2.0.3
numpy==1.24.3
imagehash==4.3.1
scikit-image==0.21.0